add Round number to decrees
This change allows to specify an increasing index (the Round number) for
decrees. Decided values with increasing Round number thus define a
replicated commit-log.

An outcome is that the AcceptorState no longers store the
"currentValue". This instead is the role of the `putCurrentValue` and
`getCurrentValue` functions from the AcceptorFunctions.

Note that current `acceptor` and `proposer` loops will only accept a
single Round at a time as the Round number is part of the Decree-number
comparison.